% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_add_lut.R
\name{acid_add_lut}
\alias{acid_add_lut}
\title{Create Add Acid Look Up Table}
\usage{
acid_add_lut(
  pCO2,
  salinityTarget,
  temperatureTarget,
  acidConc,
  targetAlk,
  minVolRatio = 2,
  maxVolRatio = 10,
  byVolRatio = 2,
  minInitPH = 7.5,
  maxInitPH = 8.5,
  byInitPH = 0.25,
  minInitSal = 20,
  maxInitSal = 40,
  byInitSal = 10,
  minInitTemp = 15,
  maxInitTemp = 30,
  byInitTemp = 5
)
}
\arguments{
\item{pCO2}{Seawater pCO2 (uatm)}

\item{salinityTarget}{Salinity target of final solution (psu)}

\item{temperatureTarget}{Temperature target for final solution (degrees C)}

\item{acidConc}{HCl concentration (Normality)}

\item{targetAlk}{Alkalinity target for the final solution (umol/kg)}

\item{minVolRatio}{LUT parameter. min(volume of final solution / initial volume)}

\item{maxVolRatio}{LUT parameter. max(volume of final solution / initial volume)}

\item{byVolRatio}{LUT parameter. Table increment for volume ratio.}

\item{minInitPH}{LUT parameter. min(initial solution pH)}

\item{maxInitPH}{LUT parameter. max(initial solution pH)}

\item{byInitPH}{LUT parameter. Table increment for initial pH}

\item{minInitSal}{LUT parameter. min(initial solution salinity). psu.}

\item{maxInitSal}{LUT parameter. max(initial solution salinity). psu.}

\item{byInitSal}{LUT parameter. Table increment for initial salinity}

\item{minInitTemp}{LUT parameter. min(initial solution temperature). degrees C}

\item{maxInitTemp}{LUT parameter. max(initial solution temperature). degrees C}

\item{byInitTemp}{LUT parameter. Table increment for initial temperature.}
}
\value{

}
\description{
Create Add Acid Look Up Table
}
