% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAcid.R
\name{addAcid}
\alias{addAcid}
\title{function to determine the amount of acid to add to reach target tank
conditions}
\usage{
addAcid(
  initVol,
  initS,
  initT,
  initPH,
  targetVol,
  targetS,
  targetT,
  targetAlk,
  acidConc,
  pCO2,
  highSwater
)
}
\arguments{
\item{initVol}{intial volume of water in the tank (L)}

\item{initS}{intial salinity of water in the tank (psu)}

\item{initT}{intial temperature of water in the tank (degrees C)}

\item{initPH}{intial pH of water in the tank}

\item{targetVol}{target volume of water in the tank (L)}

\item{targetS}{target salinity (psu)}

\item{targetT}{target temperature (degrees C)}

\item{targetAlk}{target alkalinity (umol/kg)}

\item{acidConc}{acid concentration (Normality)}

\item{pCO2}{atmospheric pCO2 (uatm)}

\item{highSwater}{salinity of the high salinty water to be added (psu)}
}
\value{
A list of the amount of acid to add and the final target pH #'
\itemize{
\item totalAcidAddInit - total vol of acid needed to adjust only
the inital water in the tank (ml)
\item aaNewTotal - total vol of acid
needed to adjust only the new water added to the tanks (ml)
\item totalAcid
\itemize{
\item total volume of acid to add (= totalAcidAddInit + aaNewTotal) (ml) note:
any of the acid additions could be negative (indcating the need to add
base), however, the need for base should not occur when using instant ocean
to create west coast water
\item finalPH - the target pH for the final
solution (this will the the pH feedback setting) }
}
}
\description{
starts with a specified volume of water with specified chemistry
conditions in then tank then calculates the amount of acid to add to reach
target volume and chemistry
}
