% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addHighSalwater.R
\name{addHighSalwater}
\alias{addHighSalwater}
\title{function to calculate the volume of high salinity water to add to head tank}
\usage{
addHighSalwater(initVol, initS, targetVol, targetS, highSwater)
}
\arguments{
\item{initVol}{intial volume of water in the  head tank (L)}

\item{initS}{intial salinity of water in the  head tank (ppt)}

\item{targetVol}{target final of water in the  head tank (L)}

\item{targetS}{target final salinitiy of water in the head tank (ppt)}

\item{highSwater}{salinity of high salinity water that will be added to the tank (ppt)}
}
\value{
A list of salt weights (g) and volumes of water to add to tank (L)
#' \itemize{
\item targetSgram - target salt weight (g) (used for internal calculation)
\item initSgram - initial salt weight (g) (used for internal calculation)
\item addSgram - salt weight to add (g) (used for internal calculation)
\item addHighSwaterVol - volume of high salinity water to add to the tank (L)
\item addDIwaterVol - volume of DI water to add to the tank (L)
}
}
\description{
function to calculate the volume of high salinity water to add to head tank
}
