% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alkIO.R
\name{alkIO}
\alias{alkIO}
\title{Instant Ocean alkaliniy from salinity with only DI water addition at 400 uatm
pCO2}
\usage{
alkIO(salinity, slope = 3200/35, intercept = 0)
}
\arguments{
\item{salinity}{Instant Ocean water salinity (psu)}

\item{slope}{empirically derived slope of Instant Ocean alklaliniy  vs
salinity relationship. Default is 3200/35.}

\item{intercept}{empirically derived slope of Instant Ocean alklaliniy  vs
salinity relationship. Default is 0.}
}
\value{
alkalinity of Instant Ocean (umol/kg)
}
\description{
this function is just a placeholder assuming a linear
relationship with alk at 30 psu = 3100 and alk at 35 psu = 3200; need to
develop an empircal curve in the range 25-60 psu
}
