% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASW-package.R
\docType{package}
\name{ASW-package}
\alias{ASW}
\alias{ASW-package}
\title{ASW: Clustering by Optimising the Average Silhouette Width}
\description{
Exact and computationally efficient algorithms for clustering by maximising the Average Silhouette Width (Rousseeuw, 1987). Provides the 'OSil' and 'FOSil' algorithms of Batool and Hennig (2021) and the 'PAMSil' algorithm of Van der Laan et al. (2003), together with 'effOSil' and 'scalOSil', which return the same clusterings as 'OSil' and 'FOSil' respectively at a fraction of the computational cost.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edelweiss611428/ASW}
  \item Report bugs at \url{https://github.com/edelweiss611428/ASW/issues}
}

}
\author{
\strong{Maintainer}: Minh Long Nguyen \email{edelweiss611428@gmail.com}

}
\keyword{internal}
