% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSil.R
\name{OSil}
\alias{OSil}
\title{Original Optimum Silhouette Clustering Algorithm}
\usage{
OSil(dist, k, initClustering = NULL, initMethod = "average")
}
\arguments{
\item{dist}{A "dist" object, which can be obtained by the "dist" function.}

\item{k}{The number of clusters.}

\item{initClustering}{A user-specified initialized clustering. It must be an integer vector of
k unique values 1,2,...,k. By defaualt, initClustering is set to NULL. If initClustering is NULL,
initMethod is used; otherwise, initClustering is used.}

\item{initMethod}{An initialization method. By default, effOSil uses average-linkage clustering
("average") for initialization. Other options include well-established clustering algorithms
such as Partition Around Medoids ("pam"), single-linkage clustering ("single"), complete-linkage
clustering ("complete"), the Ward's method ("ward.D"), etc.}
}
\value{
\describe{
\item{Clustering}{The OSil clustering solution.}
\item{ASW}{The ASW associated with the OSil clustering.}
\item{nIter}{The number of iterations needed for convergence.}
}
}
\description{
OSil implements the original Optimum Silhouette algorithm.
}
\references{
Batool F. (2019). Optimum average silhouette width clustering.
\emph{PhD Thesis}, University College London.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
