% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asw.R
\name{print.ASW}
\alias{print.ASW}
\title{Print a clustering produced by the ASW package}
\usage{
\method{print}{ASW}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ASW"}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly. Called for its side effect of printing.
}
\description{
Print a clustering produced by the ASW package
}
\examples{
effOSil(dist(scale(faithful)), K = 2:4)

}
