% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASW.R
\name{ASW}
\alias{ASW}
\title{The Average Silhouette Width}
\usage{
ASW(C, dist)
}
\arguments{
\item{C}{A clustering solution. It must be an integer vector of k unique values 1,2,...,k.}

\item{dist}{A "dist" object, which can be obtained by the "dist" function.}
}
\value{
The ASW of the clustering C with respect to the distance matrix dist.
}
\description{
This function computes the Average Silhouette Width (ASW).
}
\references{
Rousseeuw, P.J. (1987) Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. J. Comput. Appl. Math., 20, 53–65.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
