% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effOSil.R
\name{effOSil}
\alias{effOSil}
\title{The Efficient Optimum Silhouette algorithm}
\usage{
effOSil(dx, K, initMethod, variant)
}
\arguments{
\item{dx}{A "dist" object, which can be computed using stats::dist().}

\item{K}{An integer vector (or scalar) specifying the numbers of clusters. By default, K = 2:12.}

\item{initMethod}{A character vector (or string) specifying initialization methods.
By default, initMethod = "average". See ?Init for more details.}

\item{variant}{A character string specifying a variant. Options include "efficient" and "original".
If variant = "original", the original OSil algorithm is used. If variant = "efficient", effOSil is used.
By default, variant = "efficient".}
}
\value{
\describe{
\item{best_clustering}{The effOSil clustering achieving the highest ASW value.}
\item{best_asw}{The highest ASW value.}
\item{k}{The estimated number of clusters.}
\item{clusterings}{The effOSil clustering solutions for all k in K.}
\item{asw}{The ASW values associated with the effOSil clusterings.}
\item{nIter}{The numbers of iterations needed for convergence.}
}
}
\description{
This function implements the Efficient Optimum Silhouette (effOSil) algorithm.
}
\details{
This function implements the Efficient Optimum Silhouette (effOSil) algorithm, an O(N) runtime improvement of
the original, computationally expensive OSil algorithm proposed by Batool & Hennig (2021) (N is the number of
observations). An implementation of the original OSil algorithm is also available for run time comparisions.
}
\examples{
dx = dist(faithful)
effC = effOSil(dx, 2:8)
par(mfrow = c(2,1))
plot(faithful, col = effC$best_clustering, pch = 4)
plot(2:8, effC$asw, xlab = "k", ylab = "ASW")

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
