% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effOSil.R
\name{effOSil}
\alias{effOSil}
\title{The Efficient Optimum Silhouette algorithm}
\usage{
effOSil(dx, k, initClustering = NULL, initMethod = "average")
}
\arguments{
\item{dx}{A "dist" object, which can be obtained by the "dist" function.}

\item{k}{The number of clusters.}

\item{initClustering}{An initialized clustering. It must be an numeric vector of k unique values 1,2,...,k.
By default, initClustering is set to NULL. If initClustering is NULL, initMethod is used instead; otherwise, initClustering is used.}

\item{initMethod}{A character vector specifying initialization methods. It must contain only supported methods:
one of the two combined methods "multiple1" and "multiple2"; or any combination of of "pam", "average", "single",
"complete", "ward.D", "ward.D2", "mcquitty", "median", and "centroid". See ?Init for more details.}
}
\value{
\describe{
\item{Clustering}{The OSil clustering solution.}
\item{ASW}{The ASW associated with the OSil clustering.}
\item{nIter}{The number of iterations needed for convergence.}
}
}
\description{
This function implements the exact Optimum Silhouette (OSil) algorithm.
}
\details{
This function implements the exact Optimum Silhouette (OSil) algorithm proposed by Batool & Hennig (2021).
However, it is O(N) times faster than the original OSil algorithm at the cost of storing O(kN) additional values.
}
\examples{
x = iris[,-5]
dx = dist(x)
effOSil_clustering = effOSil(dx, 3, initMethod = "average")
plot(x, col = effOSil_clustering$Clustering)

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
