% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{Init}
\alias{Init}
\title{Initialization methods for Optimum Average Silhouette Width clustering algorithm}
\usage{
Init(dx, k, initMethod = "average")
}
\arguments{
\item{dx}{A "dist" object, which can be obtained by the "dist" function.}

\item{k}{The number of clusters.}

\item{initMethod}{A character vector specifying initialization methods. It must contain only supported methods:
one of the two combined methods "multiple1" and "multiple2"; or any combination of of "pam", "average", "single",
"complete", "ward.D", "ward.D2", "mcquitty", "median", and "centroid".}
}
\value{
\describe{
\item{Clustering}{An initialized clustering.}
\item{ASW}{The ASW associated with the initialized clustering.}
\item{Method}{The "best" initialization method.}
}
}
\description{
This function computes an initialized clustering for Optimum Average Silhouette Width clustering algorithms
}
\details{
This function computes an initialized clustering for Optimum Average Silhouette Width clustering algorithms
by using the clustering methods specified by initMethod, and return the clustering maximize the ASW. The two combined methods
"multiple1" and "multiple2" are:
\describe{
\item{"multiple1"}{PAM and average linkage.}
\item{"multiple2"}{PAM, average linkage, single linkage, and the Ward's method (ward.D2).}
}
}
\examples{
x = iris[,-5]
dx = dist(x)
Init(dx,3,"multiple1")

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
Batool, F., 2019. Initialization methods for optimum average silhouette width clustering. arXiv preprint arXiv:1910.08644.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
