% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Init.R
\name{Init}
\alias{Init}
\title{Initialization methods for the Optimum Silhouette algorithm}
\usage{
Init(dx, k, initMethod)
}
\arguments{
\item{dx}{dx A "dist" object, which can be computed using stats::dist().}

\item{k}{An integer scalar specifying the number of clusters.}

\item{initMethod}{A character vector (or string) specifying initialization methods. Options include any
combination of "pam", "average", "single", "complete", "ward.D", "ward.D2", "mcquitty", "median", and "centroid".
By default, initMethod = "average".}
}
\value{
\describe{
\item{clustering}{An initialized clustering.}
\item{asw}{The ASW associated with the initialized clustering.}
\item{method}{The "best" initialization method.}
}
}
\description{
This function computes an initialization for the Optimum Silhouette algorithm.
}
\details{
This function computes an initialization for the Optimum Silhouette algorithm, but it can be used as
a stand-alone clustering method.
}
\examples{
x = faithful
dx = dist(x)
Initres = Init(dx, 2, c("pam", "average", "complete"))
plot(x, col = Initres$clustering, pch = 4)
print(paste(Initres$method, "achieves the highest ASW value"))

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
Batool, F., 2019. Initialization methods for optimum average silhouette width clustering. arXiv preprint arXiv:1910.08644.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
