% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMSil.R
\name{PAMSil}
\alias{PAMSil}
\title{The PAMSil algorithm}
\usage{
PAMSil(dx, K = 2:12)
}
\arguments{
\item{dx}{A \code{dist} object, as returned by \code{\link[stats:dist]{stats::dist()}}.}

\item{K}{An integer vector specifying the numbers of clusters to consider.
Defaults to \code{2:12}.}
}
\value{
An object of class \code{"ASW"}, a list with components:
\describe{
\item{best_clustering}{An integer vector giving the clustering that
achieves the highest ASW.}
\item{best_asw}{The highest ASW value.}
\item{best_medoids}{The medoids of the clustering maximising the ASW.}
\item{k}{The estimated number of clusters.}
\item{clusterings}{An integer matrix with one column per element of \code{K}.}
\item{asw}{A numeric vector of ASW values, one per element of \code{K}.}
\item{medoids}{A list of medoid index vectors, one per element of \code{K}.}
\item{nIter}{An integer vector of iteration counts to convergence.}
\item{method}{The name of the algorithm used.}
\item{call}{The matched call.}
}
}
\description{
Clusters a distance matrix by maximising the Average Silhouette
Width (ASW) over partitions around medoids, using the PAMSil algorithm of
Van der Laan, Pollard & Bryan (2003).
}
\details{
PAMSil is a k-medoid algorithm whose objective function is the ASW.
It is initialised with \code{\link[cluster:pam]{cluster::pam()}} and then repeatedly replaces the
medoid whose replacement increases the ASW most, stopping when no replacement
improves it. Because it searches over medoid assignments rather than over
arbitrary partitions, it is more constrained than \link{effOSil} and will usually
attain a lower ASW.
}
\examples{
x = scale(faithful)
dx = dist(x)
fit = PAMSil(dx = dx, K = 2:6)

oldpar = par(mfrow = c(1, 2))
plot(faithful, col = fit$best_clustering, pch = fit$best_clustering)
plot(2:8, fit$asw, type = "l", xlab = "k", ylab = "ASW")
par(oldpar)

}
\references{
Van der Laan, M., Pollard, K. and Bryan, J. (2003). A new partitioning around
medoids algorithm. \emph{Journal of Statistical Computation and Simulation},
73(8), 575-584. \doi{10.1080/0094965031000136012}
}
\seealso{
\link{effOSil}, \link{scalOSil}, \link{asw}, \link{Silhouette}.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
