% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMSil.R
\name{PAMSil}
\alias{PAMSil}
\title{The PAMSil algorithm}
\usage{
PAMSil(dx, K)
}
\arguments{
\item{dx}{A "dist" object, which can be computed using stats::dist().}

\item{K}{An integer vector (or scalar) specifying the numbers of clusters. By default, K = 2:12.}
}
\value{
\describe{
\item{best_clustering}{The PAMSil clustering achieving the highest ASW value.}
\item{best_asw}{The highest ASW value.}
\item{best_medoids}{The medoids associated with the clustering maximize the ASW.}
\item{k}{The estimated number of clusters.}
\item{clusterings}{The PAMSil clustering solutions for all k in K.}
\item{asw}{The ASW values associated with the PAMSil clusterings.}
\item{medoids}{The medoids associated with the clustering solutions.}
\item{nIter}{The numbers of iterations needed for convergence.}
}
}
\description{
This function implements the PAMSil algorithm.
}
\details{
This function implements the PAMSil algorithm proposed by Van der Laan et al. (2003).
It is a k-medoids clustering algorithm whose objective function is the Average Silhouette Width.
}
\examples{
dx = dist(faithful)
pamsilC = PAMSil(dx, 2:8)
par(mfrow = c(2,1))
plot(faithful, col = pamsilC$best_clustering, pch = 4)
plot(2:8, pamsilC$asw, type = "l", xlab = "k", ylab = "ASW")

}
\references{
Van der Laan, M., Pollard, K. and Bryan, J., 2003. A new partitioning around medoids algorithm. Journal of Statistical Computation and Simulation, 73(8), pp.575-584.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
