% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalOSil.R
\name{scalOSil}
\alias{scalOSil}
\title{The Scalable Optimum Silhouette algorithm}
\usage{
scalOSil(dx, k, n = "default", rep = 5, initMethod = "average")
}
\arguments{
\item{dx}{A "dist" object, which can be obtained by the "dist" function.}

\item{k}{The number of clusters.}

\item{n}{The sample size. By default, n = ceiling(0.2*N).}

\item{rep}{The number of scalOSil instances. By default, rep = 5.}

\item{initMethod}{A character vector specifying initialization methods. It must contain only supported methods:
one of the two combined methods "multiple1" and "multiple2"; or any combination of of "pam", "average", "single",
"complete", "ward.D", "ward.D2", "mcquitty", "median", and "centroid". See ?Init for more details.}
}
\value{
\describe{
\item{Clustering}{Final clustering.}
\item{ASW}{The ASW of the scalOSil clustering w.r.t. dx.}
}
}
\description{
This function implements the Scalable Optimum Silhouette (scalOSil) algorithm.
}
\details{
The scalOSil algorithm is an approximation algorithm of effOSil based on subsetting. It is an improved version of FOSil (Batool & Hennig 2021).
Both the algorithms consists of two steps: partial clustering (PC) and classification (C).

In the PC-step of scalOSil, effOSil is applied to a random subset of the dataset, obtaining a subset S and its effOSil
clustering C_S. In the C-step of scalOSil, each unassigned data point is classified into one of the clusters in C_S in such
a way that the ASW is maximized.

Unlike FOSil, scalOSil runs many instances, specified by the parameter "rep", and for each instance, scalOSil only runs the
PC-step once. Moreover, the PC-step of scalOSil scales quadratically in n, while that of FOSil scales cubically in n, and
the C-step of scalOSil scales linearly in n, while that of FOSil scales quadratically in n. These allow scalOSil to handle
much larger datasets.
}
\examples{
x = iris[,-5]
dx = dist(x)
scalOSil_clustering = scalOSil(dx, 3, initMethod = "average")
plot(x, col = scalOSil_clustering$Clustering)

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
