% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalOSil.R
\name{scalOSil}
\alias{scalOSil}
\title{The Scalable Optimum Silhouette algorithm}
\usage{
scalOSil(dx, K, n, ns, rep, initMethod, variant)
}
\arguments{
\item{dx}{A "dist" object, which can be computed using stats::dist().}

\item{K}{An integer vector (or scalar) specifying the numbers of clusters. By default, K = 2:12.}

\item{n}{An integer specifying the sample size. If not specified (NULL), n is set to 0.2*N where
N is the number of observations.}

\item{ns}{An integer specifying the number of random samples used in each instance. By default, ns = 1.}

\item{rep}{An integer specifying the number of scalOSil instances. By default, rep = 10.}

\item{initMethod}{A character vector (or string) specifying initialization methods. By default,
initMethod = "average". See ?Init for more details.}

\item{variant}{A character string specifying a variant. Options include "scalable" and "original".
If variant = "original", the original FOSil algorithm is used. If variant = "scalable", scalOSil is used.
By default, variant = "scalable".}
}
\value{
\describe{
\item{best_clustering}{The scalOSil clustering achieving the highest ASW value.}
\item{best_asw}{The highest ASW value.}
\item{k}{The estimated number of clusters.}
\item{clusterings}{The scalOSil clustering solutions for all k in K.}
\item{asw}{The ASW values associated with the scalOSil clusterings.}
}
}
\description{
This function implements the Scalable Optimum Silhouette algorithm.
}
\details{
This function implements the Scalable Optimum Silhouette (scalOSil) algorithm, an O(n) runtime improvement of
the original, computationally expensive Fast OSil (FOSil) algorithm proposed by Batool & Hennig (2021) (n is
the sample size). An implementation of the original FOSil algorithm is also available for run time comparisions.
}
\examples{
dx = dist(faithful)
scalC = scalOSil(dx, 2:8)
par(mfrow = c(2,1))
plot(faithful, col = scalC$best_clustering, pch = 4)
plot(2:8, scalC$asw, xlab = "k", ylab = "ASW")

}
\references{
Batool, F. and Hennig, C., 2021. Clustering with the average silhouette width. Computational Statistics & Data Analysis, 158, p.107190.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
