% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Silhouette.R
\name{Silhouette}
\alias{Silhouette}
\title{Silhouette Width}
\usage{
Silhouette(C, dx)
}
\arguments{
\item{C}{A clustering solution. It must be an integer vector of k unique values 1,2,...,k.}

\item{dx}{A "dist" object, which can be obtained by the "dist" function.}
}
\value{
A numeric matrix of class "silhouette" containing three columns
\describe{
\item{cluster}{A clustering of the dataset.}
\item{neighbor}{The clustering labels of the nearest clusters for all data points.}
\item{sil_width}{The silhouette widths of data points.}
}
}
\description{
This function computes the Silhouette Widths of all data points in the dataset.
}
\examples{
library("cluster")
x = iris[,-5]
dx = dist(x)
C = pam(dx, 3)$clustering
Silhouette(C,dx)

}
\references{
Rousseeuw, P.J. (1987) Silhouettes: A graphical aid to the interpretation and
validation of cluster analysis. J. Comput. Appl. Math., 20, 53–65.
}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
